# Optimize comparison performance, and remove unused assigned.